Format cleanup of PHP wrapper, add --verbose flag
Prior to this PR, the PHP5 wrapper did not use standard [PHPDoc](http://www.phpdoc.de/) above the class and its methods. This made for a difficult, and often incomplete, inline documentation. Additionally, spaces and tabs were intermixed throughout the code. As spaces were the prevalent choice, this PR moves all methods to use spaces for indention. This PR also adds support for the `--verbose` command line flag, as it was previously unsupported.